marangatu-listo/requirements.txt
streamlit==1.31.0
supabase==2.3.0
python-dotenv==1.0.0
Pillow==10.2.0
google-genai==0.2.2
streamlit-camera-input-live==0.2.0
pandas==2.2.0
plotly==5.18.0
